+++ tests/e2e/alpha/internal/composition/test_download_files.py
@@ -1,4 +1,3 @@
-import itertools as it
 import os
 import pathlib
 import tempfile
@@ -10,68 +9,15 @@ from neptune_fetcher.alpha.filters import (
     Filter,
 )
 from neptune_fetcher.alpha.internal.composition.download_files import download_files
-from neptune_fetcher.alpha.internal.identifiers import RunIdentifier
 from neptune_fetcher.alpha.internal.retrieval.attribute_definitions import AttributeDefinition
-from neptune_fetcher.alpha.internal.retrieval.search import (
-    ContainerType,
-    fetch_experiment_sys_attrs,
+from neptune_fetcher.alpha.internal.retrieval.search import ContainerType
+from tests.e2e.alpha.internal.data import (
+    PATH,
+    TEST_DATA,
 )
 
 NEPTUNE_PROJECT = os.getenv("NEPTUNE_E2E_PROJECT")
-TEST_DATA_VERSION = "2025-04-19"
-EXPERIMENT_NAME = f"pye2e-fetcher-test-internal-composition-download-files-{TEST_DATA_VERSION}"
-PATH = f"test/test-internal-composition-download-files-{TEST_DATA_VERSION}"
-
-
-@pytest.fixture(scope="module")
-def run_with_attributes(client, project):
-    import uuid
-
-    from neptune_scale import Run
-
-    from neptune_fetcher.alpha.internal import identifiers
-
-    project_identifier = project.project_identifier
-
-    existing = next(
-        fetch_experiment_sys_attrs(
-            client,
-            identifiers.ProjectIdentifier(project_identifier),
-            Filter.name_in(EXPERIMENT_NAME),
-        )
-    )
-    if existing.items:
-        return
-
-    run_id = str(uuid.uuid4())
-
-    run = Run(
-        project=project_identifier,
-        run_id=run_id,
-        experiment_name=EXPERIMENT_NAME,
-    )
-
-    run.assign_files({f"{PATH}/files/file-value": b"Hello world!"})
-
-    run.close()
-
-    return run
-
-
-@pytest.fixture(scope="module")
-def experiment_identifier(client, project, run_with_attributes) -> RunIdentifier:
-    from neptune_fetcher.alpha.filters import Filter
-    from neptune_fetcher.alpha.internal.retrieval.search import fetch_experiment_sys_attrs
-
-    project_identifier = project.project_identifier
-
-    experiment_filter = Filter.name_in(EXPERIMENT_NAME)
-    experiment_attrs = _extract_pages(
-        fetch_experiment_sys_attrs(client, project_identifier=project_identifier, filter_=experiment_filter)
-    )
-    sys_id = experiment_attrs[0].sys_id
-
-    return RunIdentifier(project_identifier, sys_id)
+EXPERIMENT_NAME = TEST_DATA.experiment_names[0]
 
 
 @pytest.fixture
@@ -84,7 +30,7 @@ def test_download_files(client, project, experiment_identifier, temp_dir):
     # when
     results = download_files(
         filter_=Filter.name_in(EXPERIMENT_NAME),
-        attributes=AttributeFilter(name_eq=f"{PATH}/files/file-value"),
+        attributes=AttributeFilter(name_eq=f"{PATH}/files/file-value.txt"),
         destination=temp_dir,
         context=None,
         container_type=ContainerType.EXPERIMENT,
@@ -94,15 +40,11 @@ def test_download_files(client, project, experiment_identifier, temp_dir):
     assert len(results) == 1
     assert results[0] == (
         experiment_identifier,
-        AttributeDefinition(f"{PATH}/files/file-value", "file"),
-        temp_dir / EXPERIMENT_NAME / f"{PATH}/files/file-value",
+        AttributeDefinition(f"{PATH}/files/file-value.txt", "file"),
+        temp_dir / EXPERIMENT_NAME / f"{PATH}/files/file-value_txt",
     )
     target_path = results[0][2]
     assert target_path.exists()
     with open(target_path, "rb") as file:
         content = file.read()
         assert content == b"Hello world!"
-
-
-def _extract_pages(generator):
-    return list(it.chain.from_iterable(i.items for i in generator))
